Move rawdraw files to subdirectory.
In preparation to possibly use rawdraw as submodule.
Also changed Makefiles and includes.
