+++ ping.c
@@ -9,7 +9,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include "ping.h"
-#include "os_generic.h"
+#include "rawdraw/os_generic.h"
 
 
 #include "error_handling.h"
