+++ rawdraw/CNFGFunctions.c
