+++ rawdraw/CNFGFunctions.h
