+++ rawdraw/CNFGWinDriver.c
