+++ rawdraw/os_generic.c
