+++ searchnet.c
@@ -5,7 +5,7 @@
 #include <arpa/inet.h>
 #include <stdint.h>
 #include <stdlib.h>
-#include "os_generic.h"
+#include "rawdraw/os_generic.h"
 
 uint32_t my_random_key;
 uint8_t send_id[4];
